@@ -1,6 +1,6 @@
 module github.com/sparetimecoders/goamqp
 
-go 1.18
+go 1.20
 
 require (
 	github.com/google/uuid v1.3.0
